Added `NSObject.valuesForKeyPath`.
